Split views and program logic to separate files
In the views.py file, the code knows how to get data and the format in
which the data must be stored. The data format is checked and errors are
returned when it does not match the pattern.

In the certificator.py file, the code does not know anything about source
of the data. On the other hand it knows exactly what to do with the
data. e.g. it does not know anythong about Flask.